Support for header checks.
Added header checks to the spider. This introduces the ``naked``
parameter which can be used to indicate that a site is not using a front
end web server and thus won't have the necessary headers (and
consequently header checks should not be performed).

The format of the reports provided has also been revamped. And there is
now a ``REPORT`` file created that contains the body of the email that
is (or would be) sent.